+++ tests/unit_tests/authentication/test_authentication.py
@@ -10,7 +10,7 @@ from verda.exceptions import APIException
 INVALID_REQUEST = 'invalid_request'
 INVALID_REQUEST_MESSAGE = 'Your existence is invalid'
 
-BASE_URL = 'https://api-testing.datacrunch.io/v1'
+BASE_URL = 'https://api.example.com/v1'
 CLIENT_ID = '0123456789xyz'
 CLIENT_SECRET = 'zyx987654321'