+++ tests/unit_tests/conftest.py
@@ -4,7 +4,7 @@ import pytest
 
 from verda.http_client.http_client import HTTPClient
 
-BASE_URL = 'https://api-testing.datacrunch.io/v1'
+BASE_URL = 'https://api.example.com/v1'
 ACCESS_TOKEN = 'test-token'
 CLIENT_ID = '0123456789xyz'
 CLIENT_SECRET = '0123456789xyz'